demo/pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [{include = "gmail_assistant"}]

[tool.poetry.dependencies]
python = "^3.10"
yarl = "^1.8.2"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
fastapi = "^0.109.2"
pydantic-settings = "^2.1.0"
convore-sdk = { path = "../convore_sdk/", develop = false }
fastapi-utilities = "^0.2.0"
openai = "^1.12.0"
jinja2 = "^3.1.3"
python-dotenv = "^1.0.1"
devtools = "^0.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
